model-Tesla-Model-3/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-vertexai
google-cloud-aiplatform
pydantic
pypdf
langchain-huggingface
langchain-community
chromadb